Fix capublic update scripts bug
The capublic update scripts exited without populating capublic because
it would attempt to go to a file or directory that wasn't found, even
though it was in that directory. The exact error is as follows:

line 35: opengov_load/*: No such file or directory

I've tested this on a test bash script, and the fix seems promising,
but will keep an eye out in the following days.